+++ src/data_access/data_read.cpp
@@ -37,7 +37,7 @@ TreeErr_t TreeReadData(Tree_t* tree, const char* data_file_path)
     }
 
     TreeErr_t err = TREE_SUCCESS;
-    long int i = 0;
+    ssize_t i = 0;
 
     if ((err = ReadNode(tree, &tree->dummy->right, buffer, &i, tree->dummy)))
     {
@@ -54,7 +54,7 @@ TreeErr_t TreeReadData(Tree_t* tree, const char* data_file_path)
 
 //------------------------------------------------------------------------------------------
 
-TreeErr_t ReadNode(Tree_t* tree, TreeNode_t** node, char* buffer, long int* pos, TreeNode_t* parent)
+TreeErr_t ReadNode(Tree_t* tree, TreeNode_t** node, char* buffer, long long int* pos, TreeNode_t* parent)
 {
     assert(buffer != NULL);
     assert(tree   != NULL);
@@ -115,7 +115,7 @@ TreeErr_t ReadNode(Tree_t* tree, TreeNode_t** node, char* buffer, long int* pos,
 
 //------------------------------------------------------------------------------------------
 
-TreeErr_t ReadNodeData(char* buffer, long int* pos, char** node_data)
+TreeErr_t ReadNodeData(char* buffer, long long int* pos, char** node_data)
 {
     assert(node_data != NULL);
     assert(buffer    != NULL);
@@ -144,7 +144,7 @@ TreeErr_t ReadNodeData(char* buffer, long int* pos, char** node_data)
 
 //------------------------------------------------------------------------------------------
 
-TreeErr_t ReadNodeDataStrChr(char* buffer, long int* pos, char** node_data)
+TreeErr_t ReadNodeDataStrChr(char* buffer, long long int* pos, char** node_data)
 {
     assert(node_data != NULL);
     assert(buffer    != NULL);
@@ -181,7 +181,7 @@ TreeErr_t ReadNodeDataStrChr(char* buffer, long int* pos, char** node_data)
 
 //------------------------------------------------------------------------------------------
 
-void SkipSpaces(char* buffer, long int* pos)
+void SkipSpaces(char* buffer, long long int* pos)
 {
     assert(buffer != NULL);
     assert(pos    != NULL);
@@ -196,7 +196,7 @@ void SkipSpaces(char* buffer, long int* pos)
 
 //------------------------------------------------------------------------------------------
 
-int SkipLetter(char* buffer, long int* pos, char letter)
+int SkipLetter(char* buffer, long long int* pos, char letter)
 {
     assert(buffer != NULL);
     assert(pos    != NULL);
